+++ Source/OxyPlot/PlotModel/PlotModel.Rendering.cs
@@ -42,6 +42,8 @@ namespace OxyPlot
         {
             lock (this.SyncRoot)
             {
+                var initialClipCount = rc.ClipCount;
+
                 try
                 {
                     if (this.lastPlotException != null)
@@ -109,9 +111,19 @@ namespace OxyPlot
                     {
                         this.RenderLegends(rc);
                     }
+
+                    if (rc.ClipCount != initialClipCount)
+                    {
+                        throw new InvalidOperationException("Unbalanced calls to IRenderContext.PushClip were made during rendering.");
+                    }
                 }
                 catch (Exception exception)
                 {
+                    while (rc.ClipCount > initialClipCount)
+                    {
+                        rc.PopClip();
+                    }
+
                     // An exception was raised during rendering. This should not happen...
                     var errorMessage = string.Format(
                             "An exception of type {0} was thrown when rendering the plot model.\r\n{1}",