+++ Source/OxyPlot/Series/ContourSeries.cs
@@ -277,7 +277,7 @@ namespace OxyPlot.Series
             this.VerifyAxes();
 
             var clippingRect = this.GetClippingRect();
-            rc.SetClip(clippingRect);
+            using var _ = rc.AutoResetClip(clippingRect);
 
             var contourLabels = new List<ContourLabel>();
             var dashArray = this.LineStyle.GetDashArray();
@@ -376,8 +376,6 @@ namespace OxyPlot.Series
             {
                 this.RenderLabel(rc, cl);
             }
-
-            rc.ResetClip();
         }
 
         /// <summary>